requirements.txt
arxiv==1.4.7
poe-api
wxpusher